+++ packages/gluestack-core/CHANGELOG.md
@@ -1,10 +1,60 @@
 # @gluestack-ui-nightly/core
 
-## 1.0.10
+## 1.2.7
 
 ### Patch Changes
 
-- 25f0909: chore: testing Github Actions Workflow
+- resolved image types
+
+## 1.2.7
+
+### Patch Changes
+
+- bd7a48d: chore:test
+
+## 1.2.6
+
+### Patch Changes
+
+- added barrel packages
+- Updated dependencies
+  - @gluestack-ui-nightly/utils@1.0.20
+
+## 1.2.5
+
+### Patch Changes
+
+- created the barrel files
+- Updated dependencies
+  - @gluestack-ui-nightly/utils@1.0.19
+
+## 1.2.4
+
+### Patch Changes
+
+- add the react native imports
+
+## 1.2.3
+
+### Patch Changes
+
+- change node back to bundler
+
+## 1.2.2
+
+### Patch Changes
+
+- put type versions back
+
+## 1.2.1
+
+### Patch Changes
+
+- change to bundler
+
+## 1.1.0
+
+### Minor Changes
 
 ## 1.0.9